ethereum/requirements.txt
aptos_sdk
python-dotenv
pathlib
pyyaml
toml
requests
eth-brownie
aptos-brownie
ccxt
